Create functions to load environment vars
Create load_env_from_path function to load the .env from a specified
path. It returns the environment variables as a dictionary. If there is
an error, it will return an empty dictionary

Create load_env function to load the .env from a specified path. If
successful, it will load the environment variables into the os
environment. They should be accessed using the module os functions. If
unsuccessful, nothing happens