Create functions to make API Calls
Create get_data which makes a GET request to specified url. It requires
the arguments: url which is the url of the API you are calling. It will
raise an exception if the url is not supplied. If response has a status
code of 200, it will return data. If not, it will return an empty
dictionary

Create get_data_basic_auth which makes a GET request to specified url
using Basic Authentication. It requires the following arguments: url,
username, password. It will raise an exception if the arguments are not
supplied. It returns data if the response has a status code of 200. if
not, it will return an empty dictionary.