Add REST API calls to patch-notify.py
Implemented Changes:

Create function initialize_context. It requires as an argument: path
which is the path to the .env file or any file which supplies the ENV
variables needed for program function. It returns the env as a
dictionary if it is successful. If it is not or a empty dictionary is
received, the program will print a messages and will end prematurely.

Create function retrieve_hosts. It requires as an argument: context
which represents the ENV variables which are needed for functioning.
The function makes a call to the REST API section which supplies the
hosts information and returns the data as json if successful. If not, it
will exit the program prematurely.

Create function retrieve_hosts_collections. It requires as an argument:
context which represents the ENV variables which are needed for
functioning. The function makes a call to the REST API section which
supplies the host_collection information and returns the data as json if
successful. If not, it will exit the program prematurely. Once
successful, it will make calls to REST API for each collection to get
the detailed host collection information which includes the hosts id
associated with the collection. Once returned, the information is
appended to a list and a dictionary with this new list is returned. If
it is unsuccesful at any moment, the program will end prematurely.

Possible Changes:
- Inline Documentation
- Error Reporting