Create parse.py + functions
Create parse.py. It will house all the functions related to the
modification and sorting of data for use before email notifications

Create function reformat_structure. It requires the parameters: hosts,
host_coll which respectively are the JSON dataset retrieved from the
application instance. High Level, its function is reformat the datasets
into a singular dataset with only the required information for
distribution by email. The function returns a sorted list by host id
of host data.

Create function output_json which basically creates a .json file in the
utils directory for data provided to it.

To Add:
- Inline Documentation