Create additional functions, parse.py
Create generate_host_map function which houses the code which generates
the map that maps a specified key to the index of the item in a list. It
requires: lst - which is the list of items and key - which is the key to
use as the identifier. This code was modularised as it looked like it
would be needed in more than one function.

Create populate_external. It requires: data which is the retrieved data
from exteral data source, hosts which is the restructured host list. It
essentially handles pulling data from an external source to populate the
values in each host which where not populated during the restructuring.
It returns a fully populated host list ready for extrapolation.

Move output_json from where it was to closer to the top. Logically
associated the auxillary functions

Add TODOs for future development guide.