Test + Add parse functions, parse.py
Test functions reformat_structure, populate_external in parse.py.

Create function csv_to_json . It requires arguments: path which is the
path string to the external csv file to be converted. It opens the csv
and converts the data into a JSON format

Modify function output_json to just output an object into a json file.