Update email.py + Create Email Functions
Create string literal templates for the HTML messages. The custom data
will be dynamically added to the fields upon runtime.

Create functions create_SMTP_connection and close_SMTP_connection. These
functions will handle the lower level task of creating and closing the
SMTP connection to the local server.

Create function send_email. This handles the creation of the Message
object that is sent as an email as well as its transmission. It requires
the smtp connection, sender as an Address object, a list of Address
objects for the recepients as well as the cc list, subject and body of
the email to be sent.

Create function create_to_cc_recepients. This handles the creation of
the receipients listing from the "|" separated string of additional
contacts as well as creation of the cc listing from the context.

Create function format_table. This function handles the conversion of
the dataframe consisting host information to an HTML table format to be
inserted into the template. It requies a dataframe as well as the
notification type.

Create function email_owners. This is the high level function which
handles the customization of the emails for the emails to be sent out.
It requires the context of the program ,the data from extrapolate of
parse.py, email_type as a string which determines the email notification
which will be sent out and patch_schedule which is a string of the of
the host collection to send emails for if specified.